fix: 스케줄 추가 삭제 수정 (#204)
