+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * from './activities';
 export * from './dateform';
 export * from './myReservation';
 export * from './reservationInfo';
+export * from './activityForm';
